Add method to read from json file using Jackson
