--- src/main/java/com/solvd/RentalCompany/jackson/JsonReader.java
@@ -1,4 +1,0 @@
-package com.solvd.RentalCompany.jackson;
-
-public class JsonReader {
-}
