+++ src/main/java/com/solvd/RentalCompany/jackson/company.json
@@ -1,1 +1,32 @@
-{"rental_company":{"services":[{"id":4,"name":"WAT service","address":"Warsaw","phone":"+48745468787","country":"Poland"},{"id":5,"name":"FAT service","address":"Warsaw","phone":"+48520520630","country":"Poland"},{"id":6,"name":"RAT service","address":"Warsaw","phone":"+48501548748","country":"Poland"}],"car_services":[{"id":4,"car_id":1,"service_id":4,"price":1000.0,"date":"2022-11-05"},{"id":5,"car_id":2,"service_id":5,"price":2000.0,"date":"2022-10-10"},{"id":6,"car_id":3,"service_id":6,"price":3000.0,"date":"2022-12-15"}],"cars":[{"id":1,"vin":"XLM8592KM00152L56","category_id":"A","brand":"Kia","model":"Picanto","number":"RJA4198","year":2020,"day_price":100.0},{"id":2,"vin":"PKT001KW945444878","category_id":"A","brand":"Fiat","model":"500","number":"PLO7003","year":2019,"day_price":115.0},{"id":3,"vin":"LTY9003HY8521P961","category_id":"B","brand":"VolksWagen","model":"Polo","number":"PLO7003","year":2019,"day_price":130.0}],"categories":[{"id":"A","name":"mini cars","description":"Smaller than 370 cm in length.","coefficient":1.2},{"id":"B","name":"small cars","description":"Sizes range from 370 cm to 400 cm. ","coefficient":1.3}],"technical_inspections":[{"id":1,"car_id":1,"date":"2022-11-05","organization":"SGS Poland Sp. z o.o."},{"id":2,"car_id":2,"date":"2022-10-10","organization":"SGS Poland Sp. z o.o."},{"id":3,"car_id":3,"date":"2022-12-15","organization":"SGS Poland Sp. z o.o."}]}}
+{"rental_company":{
+  "services":
+  [
+    {"id":4,"name":"WAT service","address":"Warsaw","phone":"+48745468787","country":"Poland"},
+    {"id":5,"name":"FAT service","address":"Warsaw","phone":"+48520520630","country":"Poland"},
+    {"id":6,"name":"RAT service","address":"Warsaw","phone":"+48501548748","country":"Poland"}
+  ],
+  "car_services":
+  [
+    {"id":4,"car_id":1,"service_id":4,"price":1000.0,"date":"2022-11-05"},
+    {"id":5,"car_id":2,"service_id":5,"price":2000.0,"date":"2022-10-10"},
+    {"id":6,"car_id":3,"service_id":6,"price":3000.0,"date":"2022-12-15"}
+  ],
+  "cars":
+  [
+    {"id":1,"vin":"XLM8592KM00152L56","category_id":"A","brand":"Kia","model":"Picanto","number":"RJA4198","year":2020,"day_price":100.0}
+  ,{"id":2,"vin":"PKT001KW945444878","category_id":"A","brand":"Fiat","model":"500","number":"PLO7003","year":2019,"day_price":115.0}
+  ,{"id":3,"vin":"LTY9003HY8521P961","category_id":"B","brand":"VolksWagen","model":"Polo","number":"PLO7003","year":2019,"day_price":130.0}
+  ],
+  "categories":
+  [
+    {"id":"A","name":"mini cars","description":"Smaller than 370 cm in length.","coefficient":1.2},
+    {"id":"B","name":"small cars","description":"Sizes range from 370 cm to 400 cm. ","coefficient":1.3}
+  ],
+  "technical_inspections":
+  [
+    {"id":1,"car_id":1,"date":"2022-11-05","organization":"SGS Poland Sp. z o.o."},
+    {"id":2,"car_id":2,"date":"2022-10-10","organization":"SGS Poland Sp. z o.o."},
+    {"id":3,"car_id":3,"date":"2022-12-15","organization":"SGS Poland Sp. z o.o."}
+  ]
+}
+}
