+++ src/main/java/com/solvd/RentalCompany/models/Service.java
@@ -8,6 +8,7 @@ import java.util.Objects;
 @XmlRootElement(name="service")
 @XmlAccessorType(XmlAccessType.FIELD)
 @XmlType(name = "service", propOrder = {"id","name", "address", "phone", "country"})
+//@JsonRootName("service")
 public class Service {
     @XmlAttribute(required = true)
     @JsonProperty("id")
